Further simplified JavaScript generation by removing requirement that all controls register themselves within HTML. Now, the simple act of creating a DialogField object in JavaScript will properly register event handlers and associate labels without being told to do so by the HTML.
The main benefit of this refactoring is that we should be able to use
non-Medigy wicket controls in JavaScript handling if we wanted. Right now
the last remaining requirement for associating a JavaScript handler is to
mark a field with the JavaScriptGenerator.Contributor interface but we can
probably remove that soon, too.