Made changes to so that the Cached custom reference entities are initialized before we start loading entities from data files.
For now, the countries are added but the states are not yet. The states actually need a reference to the 
parent country object so I'm going to need to add more functionality to the EntityLoader so that it will do queries when the first line in the datafile indicates this somehow.